LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/scrollable-containing-block-validity-expected.txt


PASS .target 1
FAIL .target 2 assert_equals:
<div class="target" style="position-anchor: --a;" data-expected-width="200" data-expected-height="200"></div>
width expected 200 but got 100
PASS .target 3
FAIL .target 4 assert_equals:
<div class="target" style="position-area: top;" data-expected-width="100" data-expected-height="100"></div>
width expected 100 but got 200
